Move ConfigException into a separate file
