+++ mk/mk.deps
@@ -68,10 +68,10 @@ src/configcontainer.o: src/configcontainer.cpp include/configcontainer.h \
  include/strprintf.h include/strprintf.h include/utils.h \
  include/configcontainer.h include/logger.h
 src/configparser.o: src/configparser.cpp include/configparser.h \
- include/configactionhandler.h config.h include/exceptions.h \
- include/configparser.h include/logger.h include/strprintf.h \
- include/strprintf.h include/tagsouppullparser.h include/utils.h \
- include/configcontainer.h include/logger.h
+ include/configactionhandler.h config.h include/configexception.h \
+ include/exceptions.h include/configparser.h include/logger.h \
+ include/strprintf.h include/strprintf.h include/tagsouppullparser.h \
+ include/utils.h include/configcontainer.h include/logger.h
 src/configpaths.o: src/configpaths.cpp include/configpaths.h \
  include/cliargsparser.h include/logger.h config.h include/strprintf.h \
  include/globals.h include/strprintf.h include/utils.h \
@@ -86,7 +86,8 @@ src/controller.o: src/controller.cpp include/controller.h include/cache.h \
  include/filtercontainer.h include/fslock.h include/opml.h \
  include/urlreader.h include/queuemanager.h include/regexmanager.h \
  include/reloader.h include/remoteapi.h include/cliargsparser.h \
- include/colormanager.h include/configcontainer.h include/configparser.h \
+ include/colormanager.h include/configcontainer.h \
+ include/configexception.h include/configparser.h \
  include/downloadthread.h include/exception.h include/exceptions.h \
  include/feedhqapi.h include/fileurlreader.h include/globals.h \
  include/inoreaderapi.h include/itemrenderer.h include/htmlrenderer.h \
@@ -185,10 +186,11 @@ src/formaction.o: src/formaction.cpp include/formaction.h \
  include/history.h include/keymap.h include/configparser.h \
  include/configactionhandler.h include/rss.h include/configcontainer.h \
  include/matcher.h filter/FilterParser.h include/utils.h include/logger.h \
- config.h include/strprintf.h include/stflpp.h include/exceptions.h \
- include/logger.h include/strprintf.h include/utils.h include/view.h \
- include/colormanager.h include/controller.h include/cache.h \
- include/configpaths.h include/cliargsparser.h include/feedcontainer.h \
+ config.h include/strprintf.h include/stflpp.h include/configexception.h \
+ include/exceptions.h include/logger.h include/strprintf.h \
+ include/utils.h include/view.h include/colormanager.h \
+ include/controller.h include/cache.h include/configpaths.h \
+ include/cliargsparser.h include/feedcontainer.h \
  include/filtercontainer.h include/fslock.h include/opml.h \
  include/urlreader.h include/queuemanager.h include/regexmanager.h \
  include/reloader.h include/remoteapi.h include/filebrowserformaction.h \
@@ -335,9 +337,9 @@ src/pbcontroller.o: src/pbcontroller.cpp include/pbcontroller.h \
  include/configcontainer.h include/configparser.h \
  include/configactionhandler.h include/download.h include/fslock.h \
  include/queueloader.h include/pbcontroller.h include/colormanager.h \
- config.h include/configcontainer.h include/exceptions.h \
- include/globals.h include/keymap.h include/logger.h include/strprintf.h \
- include/nullconfigactionhandler.h include/pbview.h \
+ config.h include/configcontainer.h include/configexception.h \
+ include/exceptions.h include/globals.h include/keymap.h include/logger.h \
+ include/strprintf.h include/nullconfigactionhandler.h include/pbview.h \
  include/colormanager.h include/keymap.h include/stflpp.h \
  include/poddlthread.h include/queueloader.h include/strprintf.h \
  include/utils.h include/logger.h
