+++ src/configparser.cpp
@@ -8,6 +8,7 @@
 #include <sys/types.h>
 
 #include "config.h"
+#include "configexception.h"
 #include "exceptions.h"
 #include "logger.h"
 #include "strprintf.h"
